Agentbinary/pyproject.toml
[tool.poetry]
name = "agentbinary"
version = "0.1.0"
description = "A FastAPI-based Agent binary project"
authors = ["Dawood <dawood@appnetwise.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.32.3"
pydantic-settings = "^2.5.2"
pytest = "^8.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
coverage = "^5.5"
pytest-cov = "^5.0.0"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pytest.ini_options]
pythonpath = "src"
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]